Actually modify code when checking caching
